requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   mcp
    #   sse-starlette
    #   starlette
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
certifi==2025.6.15
    # via
    #   httpcore
    #   httpx
click==8.2.1
    # via
    #   typer
    #   uvicorn
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via mcp
httpx-sse==0.4.1
    # via mcp
idna==3.10
    # via
    #   anyio
    #   httpx
jsonschema==4.24.0
    # via mcp
jsonschema-specifications==2025.4.1
    # via jsonschema
markdown-it-py==3.0.0
    # via rich
mcp==1.10.1
    # via mix-server (pyproject.toml)
mdurl==0.1.2
    # via markdown-it-py
numpy==2.3.1
    # via pandas
pandas==2.3.0
    # via mix-server (pyproject.toml)
pyarrow==20.0.0
    # via mix-server (pyproject.toml)
pydantic==2.11.7
    # via
    #   mcp
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via mcp
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   mcp
    #   pydantic-settings
python-multipart==0.0.20
    # via mcp
pytz==2025.2
    # via pandas
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
rich==14.0.0
    # via typer
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
shellingham==1.5.4
    # via typer
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via anyio
sse-starlette==2.3.6
    # via mcp
starlette==0.47.1
    # via mcp
typer==0.16.0
    # via mcp
typing-extensions==4.14.0
    # via
    #   pydantic
    #   pydantic-core
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
tzdata==2025.2
    # via pandas
uvicorn==0.35.0
    # via mcp
